updating tests to reflect type changes in their covered classes.

+++ RaptorMRS/src/edu/wayne/cs/raptorTest/EncounterTest.java
@@ -12,8 +12,8 @@ public class EncounterTest {
 	
 	@Test
 	public void testSetAndGetEncounterID() {
-		double expectedID = 3848574;
-		double actualID;
+		int expectedID = 3848574;
+		int actualID;
 		encounter.setEncounterID(expectedID);
 		actualID = encounter.getEncounterID();
 		assert(expectedID == actualID);
